Move rootStore out of legacy
