+++ packages/satcheljs/lib/rootStore.ts
@@ -1,4 +1,4 @@
 import { ObservableMap } from 'mobx';
-import { getGlobalContext } from '../globalContext';
+import { getGlobalContext } from './globalContext';
 
 export default getGlobalContext().rootStore;
