+++ packages/satcheljs/test/legacy/rootStoreTests.ts
@@ -1,6 +1,6 @@
 import 'jasmine';
 import { isObservableMap } from 'mobx';
-import rootStore from '../../lib/legacy/rootStore';
+import rootStore from '../../lib/rootStore';
 
 describe("rootStore", () => {
     it("is an ObservableMap", () => {
